_archive_qmoi-enhanced/requirements.txt
pytest>=7.0.0
coverage>=7.0.0
requests>=2.28.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
beautifulsoup4>=4.9.3
Pillow>=9.0.0
ffmpeg-python>=0.2.0
python-magic>=0.4.24
cryptography>=3.4.7
pyOpenSSL>=20.0.1
python-dotenv>=0.19.0
aiohttp>=3.8.0
asyncio>=3.4.3
orjson>=3.6.0
ujson>=5.4.0
msgpack>=1.0.2
pyyaml>=6.0
toml>=0.10.2
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
argon2-cffi>=21.3.0
pyjwt>=2.3.0
python-multipart>=0.0.5
email-validator>=1.1.3
pydantic>=1.9.0
fastapi>=0.68.0
uvicorn>=0.15.0
starlette>=0.14.2
websockets>=10.0
aiofiles>=0.7.0
mypy>=0.910
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2
pylint>=2.9.6
bandit>=1.7.0
safety>=1.10.3 